[client] Add functionality on order's buttons
The orders buttons can do a lot of things, such as canceling, deleting, changing status do preparing or complete. There is a check when a user try to complete an order, because once you complete it, you can't interact with the order anymore, as this button will make a call on webservices that cannot be reverted.